requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
requests>=2.31.0
zeroconf>=0.131.0
pyyaml>=6.0.1
